fix: 형식 수정 및 쿼리수정

+++ backend/src/main/java/KNU/MainServer/phase2/service/Phase2Service.java
@@ -70,7 +70,7 @@ public class Phase2Service {
         List<Query10DTO> response = query10Result.stream()
                 .map(result -> Query10DTO.from(
                         (String) result[0],
-                        (Long) result[1]))
+                        (Integer) result[1]))
                 .collect(Collectors.toUnmodifiableList());
 
         return new Query10Response(response);
